refactor(web): move FileViewer's remaining pure viewport/deploy/comment/pod/asset-path rules into the file-viewer slice
Extends the file-viewer vertical-slice decomposition (ADR 0002): moves the
preview viewport/scale math, deploy provider + share-url derivations, manual-edit
inspector style normalization, markdown source-path + code-block helpers, HTML
preview asset-path resolution, comment/pod selection geometry, and bridge
coordinate clamping out of the FileViewer.tsx orchestrator and into
features/file-viewer/{rules,constants,types}.ts, reached only through the
slice's barrel. FileViewer.tsx now imports these from the barrel instead of
declaring them locally; behavior, markup, and public exports are unchanged.

Adds 183 new rules tests and keeps all 228 existing FileViewer tests green,
pnpm --filter @open-design/web typecheck clean, and pnpm guard's
vertical-slice boundary check passing.
